fix(api/v1/auth): extract route-file exports to lib/auth/api-config.ts (CI build)
Next.js 16 forbids non-HTTP-method exports from route files. Build was
failing with: 'KEY_TTL_DAYS is not a valid Route export field.'

Move all rate-limit configs, KEY_TTL_DAYS (consolidated to single source),
OAUTH_ALLOWED_REDIRECT_ORIGINS, MAX_REDIRECT_URL_LENGTH, isAllowedRedirectOrigin,
extractStateFromAuthUrl, MAX_EMAIL_LENGTH, and hashEmail to
packages/styrby-web/src/lib/auth/api-config.ts.
Routes import from there; tests update import paths.
Also removes export from handlePost in otp/send, otp/verify, and templates routes.

Affects 4 route files + their tests + 1 new lib file + templates/route.ts.

Verification:
- pnpm --filter styrby-web build: passes (was failing on KEY_TTL_DAYS export)
- pnpm --filter styrby-web typecheck: clean
- All v1/auth/* tests pass (157/157)
