+++ packages/styrby-web/src/app/api/v1/auth/oauth/start/__tests__/route.test.ts
@@ -103,7 +103,8 @@ vi.mock('@/lib/supabase/server', () => ({
 // Import handler AFTER mocks are set up
 // ============================================================================
 
-import { POST, OAUTH_START_RATE_LIMIT, OAUTH_ALLOWED_REDIRECT_ORIGINS, MAX_REDIRECT_URL_LENGTH, isAllowedRedirectOrigin, extractStateFromAuthUrl } from '../route';
+import { POST } from '../route';
+import { OAUTH_START_RATE_LIMIT, OAUTH_ALLOWED_REDIRECT_ORIGINS, MAX_REDIRECT_URL_LENGTH, isAllowedRedirectOrigin, extractStateFromAuthUrl } from '@/lib/auth/api-config';
 import * as Sentry from '@sentry/nextjs';
 import { rateLimit } from '@/lib/rateLimit';
